Refactor GameStateSystem and fix a bug
Bug
Description: The bug was in isCurrentState(GameStates)
Reproduce: Call isCurrentState with a GameStates that has not been added
to the state.
Fix: By adding a null check the problem is avoided

The refactor removed the pointer to Gwent and instead added one to an
EventListener, this is better since it can be added without initializing
the whole game. Also moved event firing to it's own method.