Fix random unit selection from empty row
Bug
Description: When Monsters ability activates it chooses a random
row and saves one random unit from that row. The problem arises
when the row is empty and the random unit index must be between 0
and 0 (exclusive).
Reproduce: Whenever the Monster's ability is activated without any
units on the board.
Fix: Now it is randomly choosing rows and checking if they have at
least one unit and if so, stop checking and save a random unit from
the row. If all rows have been checked and no row has at least one
unit then no unit will be saved.