Add Viewport and abstraction for ShaderProgram
The Viewport makes it possible to give coordinates in viewport space to
the SpriteBatch. It converts coordinates into gl coordinates specified
by glViewport. This is a workaround since I did not get glViewport to
change. Therefore the default glViewport(-0.5f, -0.5f, 0.5f, 0.5f) is
assumed in Viewport, but it can be changed through a method call.

ShaderProgram delegates uniform loading to the shader to UniformHandler.
Because ShaderProgram is now abstract, it allows every implementation to
handle their uniform variables by themselves. The method
getAllUniformLocations is used to store the location of the needed
uniform locations after the shader has been compiled. This avoids
needing to request the uniform location everytime you want to load a
value to a uniform variable in the shader.